requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncmy==0.2.9
attrs==24.2.0
boto3==1.35.15
botocore==1.35.15
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
fastapi==0.114.0
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
motor==3.5.1
numpy==2.1.2
openai==1.51.2
openapi==2.0.0
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pymongo==4.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
